+++ ECommerceAPI/Core/ECommerceAPI.Domain/Entities/OrderItem.cs
@@ -16,7 +16,7 @@ namespace ECommerceAPI.Domain.Entities
         public int ProductId { get; set; }
         public Product Product { get; set; }
 
-        public long UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public int Piece { get; set; }
     }
 }
