+++ ECommerceAPI/Core/ECommerceAPI.Domain/Entities/Product.cs
@@ -10,8 +10,8 @@ namespace ECommerceAPI.Domain.Entities
     public class Product : BaseEntity
     {
         public string Name { get; set; }
-        public long Price { get; set; }
-        public long Stock { get; set; }
+        public decimal Price { get; set; }
+        public int Stock { get; set; }
 
     }
 }
